refactor: extract media-service to packages/application
